Simplified sign in flow & can now unlink accounts
* Removed the “success” page and just redirects to the sign in page and includes an ‘action’ parameter (like “signin_google” or “unlink_facebook”) indicating the action that was just performed.

This is an improved flow and reduce edge case errors by ensuring users are always returned to the sign in page (where is then an obvious place to put any redirection logic). It also reduces the amount of code and files needed slightly so should be easier to follow.

* Have added ‘unlink’ buttons to the sign in page and now shows the name, email address and email verified status to assist people in debugging and understanding what’s going on.

The error message page has been slightly improved too.