Signing in with Twitter or Facebook returns email!
Signing in with Twitter or with Facebook now returns an email address (just as signing in with a Google account already does; this functionality was previously avalible for Facebook but stopped working recently due to a Facebook oAuth API change).

Note for Twitter this requires going to apps.twitter.com and adding links to your Terms and Conditions and Privacy Policy under the "Settings" tab, then checking the "Request email addresses" from users box under the  "Permissions" tab (otherwise users will still be able to sign in, but the app won’t be able to know their email address).

* Resolves #30
* Resolves #29