Fix for “XMLHttpRequest is not defined”
Should fix #55. I have never seen this issue in production, but it has been reported in the wild.

This fixes the only place where there was a fall through where XMLHttpRequest could be called on the server by mistake; it could occur if a session existed on the client but it had expired.