Updating dependancies to coincide with next@5.0
Updating dependancies to co-incide with todays release of Next.js 5.

Initial observations are that next@5.0 is slightly faster to build for release, and faster in production but is now slower to build/rebuild pages in development - meaning initial page views can be slow, and changes to take to CSS can take longer to take effect.

These changes in performance do not appear to be specific to this project and are a general impact of 5.0 but I’ll see what, if anything, it’s possible to do to improve performance in development mode.

Some examples features, like the CSS/SCSS handling have not changed in this release, even though there are new examples for handling them in Next.js 5, as they way it’s are currently done in this project is superior to some of the offical examples in some repsects - such as avoiding side effects like flashing on page load while CSS loads.

Some other packages are also updated to make all dependancies are up-to-date, including Bootstrap 4 Beta to the Bootstrap 4.0 release, React and React-Dom (16.0 to 16.2).

Additionally, some packages that are no longer used as direct dependancies have been removed.